rename num -> number
