+++ gadgets/src/algorithms/crh/bowe_hopwood_pedersen.rs
@@ -19,7 +19,7 @@ use crate::{
     traits::{
         algorithms::CRHGadget,
         curves::{CompressedGroupGadget, GroupGadget},
-        utilities::{boolean::Boolean, num::Number, uint::unsigned_integer::UInt8},
+        utilities::{boolean::Boolean, number::Number, uint::unsigned_integer::UInt8},
     },
 };
 use snarkvm_algorithms::crh::{
