+++ gadgets/src/algorithms/prf/blake2s.rs
@@ -20,7 +20,7 @@ use crate::traits::{
         alloc::AllocGadget,
         boolean::Boolean,
         eq::{ConditionalEqGadget, EqGadget},
-        num::Number,
+        number::Number,
         select::CondSelectGadget,
         uint::unsigned_integer::{UInt, UInt32, UInt8},
         ToBytesGadget,
