+++ gadgets/src/algorithms/signature/schnorr.rs
@@ -21,7 +21,7 @@ use crate::traits::{
         alloc::AllocGadget,
         boolean::Boolean,
         eq::{ConditionalEqGadget, EqGadget},
-        num::Number,
+        number::Number,
         uint::unsigned_integer::UInt8,
         ToBytesGadget,
     },
