+++ gadgets/src/traits/utilities/int/int_impl.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::utilities::{boolean::Boolean, num::Number};
+use crate::utilities::{boolean::Boolean, number::Number};
 
 use std::fmt::Debug;
