+++ gadgets/src/traits/utilities/int/relational/eq.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::utilities::{boolean::Boolean, eq::EvaluateEqGadget, int::*, num::Number};
+use crate::utilities::{boolean::Boolean, eq::EvaluateEqGadget, int::*, number::Number};
 use snarkvm_fields::PrimeField;
 use snarkvm_r1cs::{errors::SynthesisError, ConstraintSystem};
