+++ gadgets/src/traits/utilities/mod.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::traits::utilities::{boolean::Boolean, num::Number, uint::unsigned_integer::UInt8};
+use crate::traits::utilities::{boolean::Boolean, number::Number, uint::unsigned_integer::UInt8};
 use snarkvm_fields::Field;
 use snarkvm_r1cs::{errors::SynthesisError, ConstraintSystem};
 
@@ -24,7 +24,7 @@ pub mod bits;
 pub mod boolean;
 pub mod eq;
 pub mod int;
-pub mod num;
+pub mod number;
 pub mod select;
 pub mod uint;
