+++ gadgets/src/traits/utilities/uint/tests/uint32.rs
@@ -18,7 +18,7 @@ use crate::utilities::{
     alloc::AllocGadget,
     arithmetic::*,
     boolean::Boolean,
-    num::Number,
+    number::Number,
     uint::{UInt, UInt32},
 };
 use snarkvm_fields::{One, Zero};
