+++ gadgets/src/traits/utilities/uint/tests/uint64.rs
@@ -18,7 +18,7 @@ use crate::utilities::{
     alloc::AllocGadget,
     arithmetic::*,
     boolean::Boolean,
-    num::Number,
+    number::Number,
     uint::{UInt, UInt64},
 };
 use snarkvm_fields::{One, Zero};
